Refactor views into sub-module
