+++ src/meshweb/views/__init__.py
@@ -1,0 +1,4 @@
+from .explorer_redirect import *
+from .index import *
+from .maintenance import *
+from .website_stats import *
