+++ src/meshweb/views/maintenance.py
@@ -10,42 +10,6 @@ from rest_framework.decorators import api_view, permission_classes
 from meshapi.permissions import HasMaintenanceModePermission
 
 
-# Home view
-@extend_schema(exclude=True)  # Don't show on docs page
-@api_view(["GET"])
-@permission_classes([permissions.AllowAny])
-def index(request: HttpRequest) -> HttpResponse:
-    template = loader.get_template("meshweb/index.html")
-    links = {
-        "Member Tools": [
-            (f"{settings.FORMS_URL}/join/", "Join Form"),
-            (settings.LOS_URL, "Line of Sight Tool"),
-            (settings.MAP_URL, "Map"),
-        ],
-        "Volunteer Tools": [
-            ("/admin", "Admin Panel"),
-            ("/api/v1/geography/whole-mesh.kml", "KML Download"),
-            ("/explorer/play", "SQL Explorer"),
-            (settings.FORMS_URL, "Other Forms"),
-        ],
-        "Developer Tools": [
-            ("https://github.com/nycmeshnet/meshdb", "Source Code"),
-            ("/api/v1/", "MeshDB Data API"),
-            ("/api-docs/swagger/", "API Documentation (Swagger)"),
-            ("/api-docs/redoc/", "API Documentation (Redoc)"),
-        ],
-    }
-    context = {"links": links}
-    return HttpResponse(template.render(context, request))
-
-
-@api_view(["GET"])
-@permission_classes([permissions.AllowAny])
-def explorer_auth_redirect(request: HttpRequest) -> HttpResponse:
-    # Auth Redirect to ensure that behavior is consistent with admin panel
-    return HttpResponseRedirect("/admin/login/?next=/explorer/")
-
-
 @permission_classes([permissions.AllowAny])
 def maintenance(request: HttpRequest) -> HttpResponse:
     if not flag_enabled("MAINTENANCE_MODE"):
